+++ src/encoder/flac_encoder.ml
@@ -42,9 +42,9 @@ let encoder flac meta =
        total_samples = None;
     }
   in
-  let buf = Buffer.create Utils.pagesize in
+  let buf = Strings.Mutable.empty () in
   let write = fun chunk ->
-    Buffer.add_bytes buf chunk
+    Strings.Mutable.add_bytes buf chunk
   in
   let cb = Flac.Encoder.get_callbacks write in
   let enc = Flac.Encoder.create ~comments p cb in
@@ -66,20 +66,18 @@ let encoder flac meta =
     let b = Audio.sub b start len in
     let b = Audio.to_array b in
     Flac.Encoder.process !enc cb b;
-    let ret = Buffer.contents buf in
-    Buffer.reset buf ;
-    ret
+    Strings.Mutable.flush buf
   in
   let stop () = 
     Flac.Encoder.finish !enc cb ;
-    Buffer.contents buf
+    Strings.Mutable.flush buf
   in
     {
      Encoder.
       insert_metadata = ( fun _ -> ()) ;
       (* Flac encoder do not support header
        * for now. It will probably never do.. *)
-      header = None ;
+      header = Strings.empty ;
       encode = encode ;
       stop = stop
     }