+++ src/encoder/gstreamer_encoder.ml
@@ -121,23 +121,23 @@ let encoder ext =
        Utils.maydo Gstreamer.App_src.end_of_stream gst.audio_src;
        Utils.maydo Gstreamer.App_src.end_of_stream gst.video_src;
        GU.flush ~log gst.bin;
-       let buf = Buffer.create Utils.pagesize in
+       let buf = Strings.Mutable.empty () in
        begin
         try
          while true do
-           Buffer.add_string buf (Gstreamer.App_sink.pull_buffer_string gst.sink)
+           Strings.Mutable.add buf (Gstreamer.App_sink.pull_buffer_string gst.sink)
          done
         with
           | Gstreamer.End_of_stream -> ()
        end; 
-       Buffer.contents buf
+       buf
       end
     else
-      ""
+      Strings.Mutable.empty ()
    in
    ignore (Gstreamer.Element.set_state gst.bin Gstreamer.Element.State_null);
    GU.flush ~log gst.bin;
-   ret
+   Strings.Mutable.to_strings ret
   in
 
   let insert_metadata m =
@@ -205,17 +205,17 @@ let encoder ext =
     (* Return result. *)
     presentation_time := Int64.add !presentation_time duration;
     if !samples = 0 then
-      ""
+      Strings.empty
     else
       let ans = Gstreamer.App_sink.pull_buffer_string gst.sink in
       decr_samples ();
-      ans
+      Strings.of_string ans
   in
 
   {
     Encoder.
     insert_metadata = insert_metadata;
-    header = None;
+    header = Strings.empty;
     encode = encode;
     stop   = stop;
   }