+++ src/outputs/hls_output.ml
@@ -401,9 +401,7 @@ class hls_output p =
       in
       let oc = self#open_out fname in
       s.hls_oc <- Some (fname, oc);
-      match s.hls_encoder.Encoder.header with
-        | Some s -> output_string oc s;
-        | None -> ()
+      Strings.iter (output_substring oc) s.hls_encoder.Encoder.header
 
     method private push_current_segment =
       List.iter (fun s -> self#close_segment s) streams;
@@ -448,7 +446,7 @@ class hls_output p =
 
     method private write_pipe s b =
       let _, oc = Utils.get_some s.hls_oc in
-      output_string oc b
+      Strings.iter (output_substring oc) b
 
     method private cleanup_segments =
       self#unlink_segment current_segment;
@@ -519,10 +517,7 @@ class hls_output p =
 
     method output_stop =
       self#toggle_state `Stop;
-      let data = List.map (fun s ->
-          s.hls_encoder.Encoder.stop ()
-        ) streams
-      in
+      let data = List.map (fun s -> s.hls_encoder.Encoder.stop ()) streams in
       self#send data
 
     method output_reset =